+++ app/(auth)/verification-required/page.tsx
@@ -1,27 +1,25 @@
 import ClientOnly from "@/app/components/ClientOnly";
 import Footer from "@/app/components/Footer";
 import MainContainer from "@/app/components/MainContainer";
-import LoginModal from "@/app/components/modals/LoginModal";
 import BottomNav from "@/app/components/navbar/main/BottomNav";
 import Navbar from "@/app/components/navbar/main/Navbar";
-import VerificationClient from "@/app/components/pages/auth/VerificationClient";
+import VerificationRequiredClient from "@/app/components/pages/auth/VerificationRequiredClient";
 
 interface IParams {
   searchParams: {
-    id?: string;
+    email?: string;
   };
 }
 
-const VerificationPage = async ({ searchParams }: IParams) => {
-  const { id } = searchParams;
+const VerificationRequiredPage = async ({ searchParams }: IParams) => {
+  const { email } = searchParams;
 
   return (
     <ClientOnly>
-      <LoginModal />
       <Navbar />
       <MainContainer>
         <div className="lg:pt-12 pt-28 pb-12 lg:pb-4">
-          <VerificationClient id={id} />
+          <VerificationRequiredClient email={email} />
         </div>
       </MainContainer>
       <Footer />
@@ -30,4 +28,4 @@ const VerificationPage = async ({ searchParams }: IParams) => {
   );
 };
 
-export default VerificationPage;
+export default VerificationRequiredPage;